+++ tests/menu/test_window_pops.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 import pytest
 from pymol import cmd
 
-# simply import the wrapper you want to test
+# simply import the wrappers you want to test
 from REvoDesign.shortcuts.wrappers.designs import (wrapped_profile_pick_design,
                                                    wrapped_pssm2csv)
 from REvoDesign.shortcuts.wrappers.esm2 import wrapped_esm1v
@@ -16,7 +16,8 @@ from REvoDesign.shortcuts.wrappers.ligand_converters import (
 from REvoDesign.shortcuts.wrappers.mutation_effect_predictors import \
     wrapped_thermompnn
 from REvoDesign.shortcuts.wrappers.represents import (
-    wrapped_color_by_mutation, wrapped_color_by_plddt, wrapped_real_sc)
+    wrapped_color_by_mutation, wrapped_color_by_plddt, wrapped_load_b_factors,
+    wrapped_real_sc)
 from REvoDesign.shortcuts.wrappers.rfdiffusion_tasks import (
     wrapped_general_rfdiffusion_task, wrapped_visualize_substrate_potentials)
 from REvoDesign.shortcuts.wrappers.rosetta_tasks import (